requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
aniso8601==10.0.1
attrs==25.3.0; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
certifi==2025.10.5; python_version >= '3.7'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cloudinary==1.44.1
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
flasgger==0.9.7.1
flask==3.0.3; python_version >= '3.8'
flask-cors==5.0.0
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.11; python_version >= '3.8'
importlib-metadata==8.5.0; python_version >= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
mako==1.3.10; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.22.0; python_version >= '3.8'
marshmallow-sqlalchemy==1.1.1; python_version >= '3.8'
mistune==3.1.4; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
passlib==1.7.4
paypalrestsdk==1.13.3
pkgutil-resolve-name==1.3.10; python_version >= '3.6'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pyopenssl==25.3.0; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.2
pyyaml==6.0.3; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
rpds-py==0.20.1; python_version >= '3.8'
sendgrid==6.12.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.44; python_version >= '3.7'
stripe==13.0.1; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.20.2; python_version >= '3.8'
